add an asking before deleting a warehouse

+++ src/main/java/com/hoshino/hoshinoscene/tools/contextMenu/ForWarehouse.java
@@ -3,10 +3,13 @@ package com.hoshino.hoshinoscene.tools.contextMenu;
 import com.hoshino.hoshinoscene.Controller;
 import com.hoshino.hoshinoscene.HoshinoApp;
 import com.hoshino.hoshinoscene.custom.WarehouseStyle;
+import javafx.scene.control.Alert;
+import javafx.scene.control.ButtonType;
 import javafx.scene.control.ContextMenu;
 import javafx.scene.control.MenuItem;
 import java.io.File;
 import java.io.IOException;
+import java.util.Optional;
 
 public class ForWarehouse extends ContextMenu {
     //public static final String ICON_STYLE = "-fx-font-size: 12";
@@ -19,16 +22,24 @@ public class ForWarehouse extends ContextMenu {
         MenuItem del = new MenuItem("删除"/*, iconOfDel*/);
         del.setOnAction(e-> {
             File file = new File("json\\warehouses\\"+ws.getId()+".json");
-            if(file.delete()){
-                System.out.println("成功删除了"+ws.getId()+".json");
-                Controller c = HoshinoApp.fxmlLoader.getController();
-                try {
-                    c.load();
-                } catch (IOException ex) {
-                    throw new RuntimeException(ex);
+            Alert delAsking = new Alert(Alert.AlertType.CONFIRMATION);
+            delAsking.setTitle("删除提醒");
+            delAsking.setHeaderText("确定要删除"+ws.getId()+".json吗");
+            Optional<ButtonType> result = delAsking.showAndWait();
+            if(result.isPresent()) {
+                if (result.get() == ButtonType.OK) {
+                    if (file.delete()) {
+                        System.out.println("成功删除了" + ws.getId() + ".json");
+                        Controller c = HoshinoApp.fxmlLoader.getController();
+                        try {
+                            c.load();
+                        } catch (IOException ex) {
+                            throw new RuntimeException(ex);
+                        }
+                    } else {
+                        System.out.println("删除失败");
+                    }
                 }
-            } else {
-                System.out.println("删除失败");
             }
         });
